chore: rename group file
