+++ .changeset/tender-nails-stick.md
@@ -1,0 +1,6 @@
+---
+"@ckb-ccc/core": patch
+---
+
+Abstract transaction completion step by a new layer named FeePayer
+
