+++ packages/core/src/signer/feePayer/feePayerGroup.ts
@@ -1,6 +1,6 @@
-import { FeePayer, FeeRateOptions } from "./index.js";
 import { Transaction, TransactionLike } from "../../ckb/transaction.js";
 import { Client } from "../../client/client.js";
+import { FeePayer, FeeRateOptions } from "./feePayer.js";
 
 export class FeePayerGroup extends FeePayer {
   constructor(private feePayers: FeePayer[]) {
@@ -24,7 +24,11 @@ export class FeePayerGroup extends FeePayer {
     return tx;
   }
 
-  async completeTxFee(tx: Transaction, client: Client, options?: FeeRateOptions): Promise<void> {
+  async completeTxFee(
+    tx: Transaction,
+    client: Client,
+    options?: FeeRateOptions,
+  ): Promise<void> {
     for (const payer of this.feePayers) {
       await payer.completeTxFee(tx, client, options);
     }
